requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17.2
amqp==5.0.6
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==20.3.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
compressed-rtf==1.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.2
discord-components==2.1.2
discord.py==1.7.3
Django==2.2.24
django-js-asset==1.2.2
django-timezone-field==4.2.1
djangorestframework==3.12.4
dlms-cosem==21.3.1
ebcdic==1.1.1
future==0.18.2
greenlet==1.1.2
idna==3.3
IMAPClient==2.1.0
inflection==0.5.1
itypes==1.2.0
Jinja==23.0.2
jsonpointer==2.1
jsonschema==3.2.0
kombu==5.1.0
MarkupSafe==2.0.1
multidict==5.2.0
olefile==0.46
orderedmultidict==1.0.1
packaging==21.0
pefile==2021.9.3
pip==21.3.1
ply==3.11
prompt-toolkit==3.0.21
pycparser==2.20
pyinstaller==4.6
pyinstaller-hooks-contrib==2021.3
pyparsing==3.0.4
pyrsistent==0.18.0
pyserial==3.5
pyTelegramBotAPI==4.1.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
PyYAML==5.4.1
regex==2021.10.23
requests==2.26.0
rfc3987==1.3.8
ruamel.yaml==0.17.16
setuptools==57.4.0
six==1.16.0
sqlparse==0.4.2
strict-rfc3339==0.7
tld==0.12.6
typing-extensions==3.7.4.3
tzdata==2021.5
tzloca==l4.1
uritemplate==4.1.1
urllib==31.26.7
validate-email==1.3
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
Whoosh==2.7.4
yar==l1.7.0
zope.event==4.5.0
zope.interface==5.4.0